clear all mocks after each test (createTree)
